endgame - Copy (2)/requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
plyer==2.1.0
pyautogui==0.9.54
wikipedia==1.4.0
pywhatkit==5.4
google-generativeai==0.3.1
PyQt5==5.15.10
python-pptx==0.6.21
Pillow==10.0.1
opencv-python==4.8.0.74
numpy==1.24.3
requests==2.31.0 